document-processor-api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==0.8.11
python-pptx==0.6.22
Werkzeug==2.3.7
openai==1.52.0
